sfilter/src/requirements.txt
transformers
torch
Flask
gunicorn
requests
google-cloud-storage==2.10.0
